MAR-922: Use the Oauth2 service to get users
We cannot use the the django User object directly.
We get the user GUID from the SSO service
The Profile object has a GUID field and a one-to-one connection with the User
object
So, we use the SSO service to get the GUID, we use the GUID to get the Profile
object, We use the Profile object to get the User object. This is done because
the SSO GUID is incompatible with the django auth system