Add a test checking that the `User` fields in the progress update have been serialised.
The problem arose because the REST framework's serialisers are quite happy to leave objects as objects if it doesn't have a serialiser for them. Then the JSON serialiser throws up its hands in dismay, but that happens deep down in the response object, so checking here seems good enough to detect the problem.